Visitor-ify ExpEmbedding validity; SAMify PurityContext
Move the validity dispatch out of the IR:

- Replace `ExpEmbedding.isValid` (one non-default override on `Assert`)
  with `ValidityVisitor : ExpVisitor<Boolean>` living next to
  `ExprPurityVisitor` in the purity package.
- `PurityContext` becomes a `fun interface`; `DefaultPurityContext` is
  deleted and constructed inline as a lambda at the sole call site.
- Move `preorder` and `expressionSource` from `core/purity/ExprUtils.kt`
  to `core/embeddings/expression/Meta.kt`, next to `WithPosition` and
  `SharingContext`. The purity package no longer owns generic
  tree-walks; `FullNamedFunctionSignature` imports `preorder` from its
  new home.

No behavior change: the only validity rule (Assert condition must be
pure) keeps its current semantics. `ExpEmbedding` no longer depends on
the purity package.
